api/pyproject.toml
[project]
name = "mimir-api"
version = "0.1.0"
description = "Mission Control dashboard API for Ullr"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "aiosqlite>=0.20.0",
    "watchdog>=6.0.0",
    "pyyaml>=6.0",
    "sse-starlette>=2.0.0",
    "ullr",
]

[tool.uv.sources]
ullr = { path = "../../ullr", editable = true }

[project.scripts]
mimir-api = "mimir_api.app:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mimir_api"]

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "A", "SIM", "TCH"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[[tool.mypy.overrides]]
module = ["ullr.*"]
ignore_missing_imports = true
